MATE-14 : [RENAME] MemberInfo 클래스 패키지 변경
